Move command into a single file
