+++ src/beyond_identity/api/identities/mod.rs
@@ -1,9 +1,3 @@
 pub mod api;
-pub mod create;
-pub mod delete;
-pub mod get;
-pub mod list;
-pub mod list_groups;
-pub mod list_roles;
-pub mod patch;
+pub mod command;
 pub mod types;
